tree-wide: use fixed-width integers for user-provided values where it makes sense
We want the same behavior accross architectures.

+++ src/launch/config.h
@@ -150,7 +150,7 @@ struct ConfigNode {
 
                 struct {
                         unsigned int name;
-                        unsigned long long value;
+                        uint64_t value;
                 } limit;
 
                 struct {
